requirements.txt
PyYAML>=3
Twisted==13.1.0
async==0.6.1
blinker==1.3
github3.py==0.9
pystache==0.5
requests>=2.2
semantic_version==2.3
GitPython==0.3.2.RC1